Проект Спринта 6 v2.0 (после ревью)

+++ yatube/posts/views.py
@@ -16,7 +16,7 @@ def index(request):
         'author',
         'group'
     )
-    page_obj = utils.paginate(request, post_list)
+    page_obj = utils.paginate_page(request, post_list)
     context = {
         'page_obj': page_obj,
     }
@@ -27,7 +27,7 @@ def group_posts(request, slug):
     """Посты группы. Применяется паджинатор."""
     group = get_object_or_404(Group, slug=slug)
     post_list = group.posts.select_related('author', 'group')
-    page_obj = utils.paginate(request, post_list)
+    page_obj = utils.paginate_page(request, post_list)
     context = {
         'group': group,
         'page_obj': page_obj,
@@ -43,7 +43,7 @@ def profile(request, username):
         user=request.user,
         author=(User.objects.get(username=username))
     ).exists()
-    page_obj = utils.paginate(request, post_list)
+    page_obj = utils.paginate_page(request, post_list)
     context = {
         'author': author,
         'page_obj': page_obj,
@@ -56,13 +56,9 @@ def post_detail(request, post_id):
     """Страница поста: вывод подробной информации о посте"""
     post = get_object_or_404(Post, id=post_id)
     form = CommentForm(request.POST or None)
-    if request.method == 'POST':
-        comments = add_comment(request, post_id)
-    comments = post.comments.select_related('post')
     context = {
         'post': post,
-        'form': form,
-        'comments': comments
+        'form': form
     }
     return render(request, 'posts/post_detail.html', context)
 
@@ -72,14 +68,12 @@ def post_create(request):
     """Создание нового поста, после успешного заполнения -
     переход на страницу профиля"""
     form = PostForm(request.POST or None, files=request.FILES or None)
-    if request.method == 'POST':
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.author = request.user
-            post.save()
-            return redirect('posts:profile', request.user)
-
-    return render(request, 'posts/post_create.html', {'form': form})
+    if request.method != 'POST' or not form.is_valid():
+        return render(request, 'posts/post_create.html', {'form': form})
+    post = form.save(commit=False)
+    post.author = request.user
+    post.save()
+    return redirect('posts:profile', request.user)
 
 
 @login_required
@@ -96,19 +90,19 @@ def post_edit(request, post_id):
     if request.user != post.author:
         return redirect('posts:post_detail', post_id)
 
-    if request.method == 'POST':
-        if form.is_valid():
-            post = form.save()
-            post.save()
-        return redirect('posts:post_detail', post_id)
-
-    return render(request,
-                  'posts/post_create.html',
-                  {'form': form,
-                   'is_edit': True,
-                   'post': post
-                   }
-                  )
+    if request.method != 'POST' or not form.is_valid():
+        return render(
+            request,
+            'posts/post_create.html',
+            {
+                'form': form,
+                'is_edit': True,
+                'post': post
+            }
+        )
+    post = form.save()
+    post.save()
+    return redirect('posts:post_detail', post_id)
 
 
 @login_required
@@ -122,7 +116,6 @@ def add_comment(request, post_id):
         comment.author = request.user
         comment.post = post
         comment.save()
-    print(form.errors)
     return redirect('posts:post_detail', post_id=post_id)
 
 
@@ -134,7 +127,7 @@ def follow_index(request):
     post_list = Post.objects.filter(
         author__following__user=user
     ).select_related('author', 'group')
-    page_obj = utils.paginate(request, post_list)
+    page_obj = utils.paginate_page(request, post_list)
     context = {
         'page_obj': page_obj,
     }
@@ -143,7 +136,8 @@ def follow_index(request):
 
 @login_required
 def profile_follow(request, username):
-    if User.objects.get(
+    if get_object_or_404(
+        User,
         username=username
     ) == request.user or Follow.objects.filter(
         user=request.user,
@@ -163,5 +157,6 @@ def profile_unfollow(request, username):
         user=request.user,
         author=(User.objects.get(username=username))
     )
-    unfollow.delete()
-    return redirect('posts:index')
+    if unfollow.exists():
+        unfollow.delete()
+        return redirect('posts:index')
